Replace "sandbox" with "safebox"!
Although the word "sandbox" is arbitrary, it was a poor choice because in computing a "sandbox" is an isolated environment where code or programs run safely without affecting the rest of the system. There is no such environment in this project; in our case, we prefer to refer to our storage image as a "safebox". A safebox is a secure, encrypted container on a computer or server designed to store sensitive data (such as passwords, confidential documents, or credit card details) safely. Only users or processes with the proper credentials can access or modify its contents.

With this in mind, I decided to change the term "sandbox" to "safebox". This is a major change and requires a version update.